+++ src/Controlador/Main.java
@@ -3,6 +3,7 @@ package Controlador;
 import javax.swing.JOptionPane;
 
 import ConexionBD.Conexion;
+import ConexionBD.DAO_Consultar;
 import Interfaz.VentanaPrincipal;
 import Modelo.Fachada;
 import ConexionBD.DAO_Crear;
@@ -12,7 +13,7 @@ import ConexionBD.DAO_Crear;
 		public static void main(String[] args) {
 			String nombreTienda;
 			Conexion conex = new Conexion();
-			DAO_Crear dao = new DAO_Crear(conex.getConnection());
+			DAO_Consultar dao = new DAO_Consultar(conex.getConnection());
 			nombreTienda=dao.ObtenerFachada();
 			Fachada.getInstance().setNombre(nombreTienda);
 			VentanaPrincipal login = new VentanaPrincipal(Fachada.tienda);
